pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hs-copilot"
version = "0.1.0"
description = "Facility-scoped doctor copilot service for HealthStack"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.34,<1.0",
  "pymongo>=4.10,<5.0",
  "pydantic-settings>=2.8,<3.0",
  "PyJWT>=2.10,<3.0",
  "python-dotenv>=1.0,<2.0",
  "streamlit>=1.43,<2.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
